UI-DESIGN: Adding new button with FS.onCreateNewProjectAndFile() function to allow a user to create a new project and save a new file within that project
The user is first prompted to enter a name for the new project AKA folder. After that is completed, a new editor is shown and the user selects a filetype. After selecting a filetype, the user is shown a dialog asking to enter a name for the file. After providing a file name and clicking OK, the new file is saved in the folder created.

Hiding the tabs.

The user is first prompted to enter a name for the new project AKA folder. After that is completed, a new editor is shown and the user selects a filetype. After selecting a filetype, the user is shown a dialog asking to enter a name for the file. After providing a file name and clicking OK, the new file is saved in the folder created.

Hiding the tabs.